Improved import_module(), moved bot_func to modules/util
