+++ modules/music.py
@@ -11,8 +11,7 @@ from subprocess import PIPE, STDOUT
 
 PATH = os.path.dirname(os.path.abspath(__file__))
 
-sys.path.insert(0, os.path.dirname(PATH))
-from bot_func import Bot
+from modules.util import Bot
 
 PERMISSION = 5
