+++ modules/nico.py
@@ -19,8 +19,7 @@ from subprocess import PIPE, DEVNULL
 
 PATH = os.path.dirname(os.path.abspath(__file__))
 
-sys.path.insert(0, os.path.dirname(PATH))
-from bot_func import Bot
+from modules.util import Bot
 
 logger = logging.getLogger()
